FIR/UAST: commonize type check expression
