+++ plugins/uast-kotlin-fir/testData/type/typeCheck.kt
@@ -1,0 +1,6 @@
+fun Any?.asString(): String {
+    return if (x !is String)
+        x.toString()
+    else
+        x
+}
